+++ schema_validator/typing.py
@@ -1,4 +1,5 @@
-from typing import Dict, List, Tuple, Type, TypedDict, Union
+from typing import Dict, List, Tuple, Type, TypedDict, Union, Optional
+from dataclasses import dataclass, field
 
 from pydantic import BaseModel
 from flask.typing import (
@@ -31,3 +32,23 @@ class ServerObject(TypedDict, total=False):
     url: str
     description: str
     variables: Dict[str, VariableObject]
+
+
+class ResponseSchema(BaseModel):
+    """
+        base response to inherit
+
+        class TodoResponse(ResponseSchema):
+            name: str
+            age: int
+    """
+    success: Optional[bool] = True
+    error_no: Optional[int] = 0
+    error_message: Optional[str] = ""
+
+
+@dataclass
+class ResponseClass:
+    success: Optional[bool] = True
+    error_no: Optional[int] = 0
+    error_message: Optional[str] = ""