+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
-0.1.7 2021-09-28
-----------------
+0.1.12 2021-09-29
+-----------------
+* the default response schema
+* support show the swagger which include the {tag}
+* support export the swagger which include the {tag}
+
+0.1.11 2021-09-28
+-----------------
+
+* bugfix
 * support tag for swagger
 
 0.1.6 2021-09-27